# src/Cvoya.Spring.Host.Api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Cvoya.Spring.Host.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/packages/{package}/templates/{name}": {
      "get": {
        "tags": [
          "Packages"
        ],
        "summary": "Returns the raw YAML for a unit template inside a package",
        "operationId": "GetUnitTemplate",
        "parameters": [
          {
            "name": "package",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitTemplateDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/packages/{name}": {
      "get": {
        "tags": [
          "Packages"
        ],
        "summary": "Returns detailed contents of a single package",
        "operationId": "GetPackage",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackageDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentTemplateSummary": {
        "required": [
          "package",
          "name",
          "displayName",
          "role",
          "description",
          "path"
        ],
        "type": "object",
        "properties": {
          "package": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "displayName": {
            "type": [
              "null",
              "string"
            ]
          },
          "role": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "path": {
            "type": "string"
          }
        }
      },
...
      "ConnectorSummary": {
        "required": [
          "package",
          "name",
          "path"
        ],
        "type": "object",
        "properties": {
          "package": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
...
      "PackageDetail": {
        "required": [
          "name",
          "description",
          "unitTemplates",
          "agentTemplates",
          "skills",
          "connectors",
          "workflows"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "unitTemplates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UnitTemplateSummary"
            }
          },
          "agentTemplates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentTemplateSummary"
            }
          },
          "skills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkillSummary"
            }
          },
          "connectors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectorSummary"
            }
          },
          "workflows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkflowSummary"
            }
          }
        }
      },
...
      "SkillSummary": {
        "required": [
          "package",
          "name",
          "hasTools",
          "path"
        ],
        "type": "object",
        "properties": {
          "package": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "hasTools": {
            "type": "boolean"
          },
          "path": {
            "type": "string"
          }
        }
      },
...
      "UnitTemplateDetail": {
        "required": [
          "package",
          "name",
          "path",
          "yaml"
        ],
        "type": "object",
        "properties": {
          "package": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "yaml": {
            "type": "string"
          }
        }
      },
      "UnitTemplateSummary": {
        "required": [
          "package",
          "name",
          "description",
          "path"
        ],
        "type": "object",
        "properties": {
          "package": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "path": {
            "type": "string"
          }
        }
      },
...
      "WorkflowSummary": {
        "required": [
          "package",
          "name",
          "path"
        ],
        "type": "object",
        "properties": {
          "package": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      }
    }
  },
...
}